Extract Codex CLI scheduler packet builders
